+++ packages/algo/bit/tests/BitGrid_test.ts
@@ -30,7 +30,7 @@ Deno.test("BitGrid construct", () => {
 Deno.test("BitGrid", () => {
   const grid = BitGrid.make({ width: 32, height: 32 });
   assertEquals(grid.getWidth(), 32);
-  assertEquals(grid.getWidth32(), 1);
+  assertEquals(grid.getInternalUint32ArrayWidth(), 1);
   assertEquals(grid.getHeight(), 32);
   assertEquals(grid.getPopulation(), 0);
   assertEquals(grid.hasAliveCellAtBorder(), false);
@@ -168,7 +168,7 @@ Deno.test("BitGrid forEach", () => {
 Deno.test("BitGrid make", () => {
   const grid = BitGrid.make({ width: 48, height: 32 });
   assertEquals(grid.getWidth(), 64);
-  assertEquals(grid.getWidth32(), 2);
+  assertEquals(grid.getInternalUint32ArrayWidth(), 2);
   assertEquals(grid.getHeight(), 32);
 });
 
@@ -194,7 +194,7 @@ Deno.test("BitGrid get set unset", () => {
   assertEquals(grid.get(2, 3), 0);
 });
 
-Deno.test("BitGrid setTo", () => {
+Deno.test("BitGrid setStateAt", () => {
   const grid = BitGrid.make({ width: 32, height: 5 });
   grid.setStateAt(2, 3, 1);
   assertEquals(grid.get(2, 3), 1);